Act gracefully when more than 6 keys are reported at once
When making a keyboard, it's easy for the user to mash plenty of
keys at once. While a keyboard firmware _could_ trap the ValueError and
continue gracefully, I think it would be nice if instead the oldest
keycode is dropped from the report to make room for the newest one.

This does slightly complicate the code but it makes all my keyboard
projects more robust.

The pylint disables are in order to avoid new memory allocations,
which enumerate() and tuple construction both do.